pyproject.toml
[project]
name = "codeseoul-discord-bot"
version = "0.0.1"
description = "stuff and things"
requires-python = ">=3.11"
dependencies = [
    "discord >=2.3.1,<3.0.0",
    "google-cloud-translate >=3.11.2,<4.0.0",
]

[project.scripts]
bot = "codeseoul_discord_bot:main.main"